+++ exporters/prometheus/test/exporter_utils_test.cc
@@ -269,16 +269,16 @@ TEST(PrometheusExporterUtils, TranslateToPrometheusHistogramNormal)
   ASSERT_EQ(checked_label_num, 3);
 }
 
-class SanitizeNameTest : public ::testing::Test
+class SanitizeTest : public ::testing::Test
 {
   Resource resource_ = Resource::Create({});
   nostd::unique_ptr<InstrumentationScope> instrumentation_scope_ =
       InstrumentationScope::Create("library_name", "1.2.0");
 
 protected:
-  void CheckSanitation(const std::string &original, const std::string &sanitized)
+  void CheckSanitizeName(const std::string &original, const std::string &sanitized)
   {
-    metric_sdk::InstrumentDescriptor instrument_descriptor_{
+    metric_sdk::InstrumentDescriptor instrument_descriptor{
         original, "description", "unit", metric_sdk::InstrumentType::kCounter,
         metric_sdk::InstrumentValueType::kDouble};
     std::vector<prometheus::MetricFamily> result = PrometheusExporterUtils::TranslateToPrometheus(
@@ -286,20 +286,45 @@ protected:
          std::vector<metric_sdk::ScopeMetrics>{
              {instrumentation_scope_.get(),
               std::vector<metric_sdk::MetricData>{
-                  {instrument_descriptor_, {}, {}, {}, {{{}, {}}}}}}}},
+                  {{instrument_descriptor, {}, {}, {}, {{{}, {}}}}}}}}},
         false);
     EXPECT_EQ(result.begin()->name, sanitized + "_unit");
   }
+
+  void CheckSanitizeLabel(const std::string &original, const std::string &sanitized)
+  {
+    metric_sdk::InstrumentDescriptor instrument_descriptor{
+        "name", "description", "unit", metric_sdk::InstrumentType::kCounter,
+        metric_sdk::InstrumentValueType::kDouble};
+    std::vector<prometheus::MetricFamily> result = PrometheusExporterUtils::TranslateToPrometheus(
+        {&resource_,
+         std::vector<metric_sdk::ScopeMetrics>{
+             {instrumentation_scope_.get(),
+              std::vector<metric_sdk::MetricData>{
+                  {instrument_descriptor, {}, {}, {}, {{{{original, "value"}}, {}}}}}}}},
+        false);
+    EXPECT_EQ(result.begin()->metric.begin()->label.begin()->name, sanitized);
+  }
 };
 
-TEST_F(SanitizeNameTest, All)
+TEST_F(SanitizeTest, Name)
 {
-  CheckSanitation("name", "name");
-  CheckSanitation("name?", "name_");
-  CheckSanitation("name???", "name_");
-  CheckSanitation("name?__", "name_");
-  CheckSanitation("name?__name", "name_name");
-  CheckSanitation("name?__name:", "name_name:");
+  CheckSanitizeName("name", "name");
+  CheckSanitizeName("name?", "name_");
+  CheckSanitizeName("name???", "name_");
+  CheckSanitizeName("name?__", "name_");
+  CheckSanitizeName("name?__name", "name_name");
+  CheckSanitizeName("name?__name:", "name_name:");
+}
+
+TEST_F(SanitizeTest, Label)
+{
+  CheckSanitizeLabel("name", "name");
+  CheckSanitizeLabel("name?", "name_");
+  CheckSanitizeLabel("name???", "name_");
+  CheckSanitizeLabel("name?__", "name_");
+  CheckSanitizeLabel("name?__name", "name_name");
+  CheckSanitizeLabel("name?__name:", "name_name_");
 }
 
 class AttributeCollisionTest : public ::testing::Test